Move DiscoverHeader to discover-stream to render it inside the stream
